Refactor

+++ app/Controllers/OrdersController.php
@@ -4,6 +4,7 @@ namespace App\Controllers;
 
 use App\Models\OrderModel;
 use App\Models\ParticipantModel;
+use App\Models\ProductModel;
 use App\Models\UserModel;
 use CodeIgniter\Controller;
 
@@ -11,12 +12,18 @@ use CodeIgniter\Controller;
 class OrdersController extends Controller
 {
   public function delete($productId)
-  {
+{
     $orderModel = new OrderModel();
+    $productModel = new ProductModel(); // Add this line
+
+    $data = [
+      'quantity' => $this->request->getVar('quantity'),
+    ];
+    
+    $quantity = $this->request->getVar('quantity');
 
     $userId = session()->get('id');
 
-    // Check if the user has a reservation for the specified event
     $order = $orderModel->where('product_id', $productId)
                         ->where('user_id', $userId)
                         ->first();
@@ -26,9 +33,16 @@ class OrdersController extends Controller
         return redirect()->back()->with('error', 'Prenotazione non trovata.');
     }
 
-    // Delete the reservation
-    $orderModel->delete($order['id']);
+    if ($quantity == $order['quantity']) {
+        // Delete the reservation
+        $orderModel->delete($order['id']);
+        return redirect()->to('/cash')->with('success', 'Prenotazione annullata.');
+    } elseif ($quantity < $order['quantity']) {
+        $newQuantity = $order['quantity'] - $quantity;
+        $data['quantity'] = $newQuantity;
+        $orderModel->update($order['id'], $data); // Use $productModel to update the quantity
+        return redirect()->to('/cash')->with('success', 'Prenotazione annullata.');
+    }
+}
 
-    return redirect()->to('/cash')->with('success', 'Prenotazione annullata.');
-  }
 }
